pyproject.toml
[project]
name = "gymapi"
version = "0.1.0"
description = ""
authors = [
    {name = "Maxim",email = "yrev2011@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "dishka (>=1.5.3,<2.0.0)",
    "fastapi[standard] (>=0.115.12,<0.116.0)",
    "pydantic-settings (>=2.8.1,<3.0.0)",
    "alembic (>=1.15.2,<2.0.0)",
    "asyncpg (>=0.30.0,<0.31.0)",
    "sqlalchemy (>=2.0.40,<3.0.0)",
]

[tool.poetry]

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.5"
faker = "^37.1.0"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
